+++ .idea/misc.xml
@@ -18,6 +18,7 @@
         <entry key="../../../../../../layout/compose-model-1659161229168.xml" value="0.13240740740740742" />
         <entry key="../../../../../../layout/compose-model-1659162495866.xml" value="0.1" />
         <entry key="../../../../../../layout/compose-model-1659183658956.xml" value="0.13240740740740742" />
+        <entry key="../../../../../../layout/compose-model-1659260736586.xml" value="0.13240740740740742" />
         <entry key="main/src/main/res/drawable/ic_another_journey_active.xml" value="0.1635" />
         <entry key="main/src/main/res/drawable/ic_another_journey_inactive.xml" value="0.1635" />
         <entry key="main/src/main/res/drawable/ic_artboard_1.xml" value="0.1" />
